Updated SHHS mappings and attribute lexicon.
Added to medical history and medications.

Three variables under medical history do not have a health interview
form linked with them:  prev_hx_mi, srhype, and prev_hx_stroke.  Based
on the MOP it appears as though this history would have been gathered
during the health interview.  I left "Health interview" as the equipment
fro these 3 variables.

I changed the equipment to "Health interview form" for the medical
history variables  that are linked with a health interview form.
I added Health interview form to the attribute lexicon, but can delete
it if the distinction is not necessary.